+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to the "Markdown Org" extension will be documented in this f
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.17.0] - 2026-08-16
 
 ### Added
 
@@ -1371,7 +1371,8 @@ and repeating tasks that move forward when marked DONE instead of closing.
 
 - Requires [markdown-org-extract](https://crates.io/crates/markdown-org-extract) utility
 
-[Unreleased]: https://github.com/VitalyOstanin/markdown-org-vscode/compare/v0.16.0...HEAD
+[Unreleased]: https://github.com/VitalyOstanin/markdown-org-vscode/compare/v0.17.0...HEAD
+[0.17.0]: https://github.com/VitalyOstanin/markdown-org-vscode/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/VitalyOstanin/markdown-org-vscode/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/VitalyOstanin/markdown-org-vscode/compare/v0.13.0...v0.15.0
 [0.13.0]: https://github.com/VitalyOstanin/markdown-org-vscode/compare/v0.12.0...v0.13.0